creation of itsModel.py
itsModel.py was created to keep all classes and functions related to the model and that can be changed by including heterogeneous rockets. Its help to keep the fine tune code part protected from alterations and make the code more intelligible.
